more fixes

+++ Classes/LoanDetails.cs
@@ -30,7 +30,7 @@ namespace PQ_API.Classes
         public DateTime ApplicationDate {get;set;}
         public DateTime DisbursementDate {get;set;}
         public Enums.PaymentFrequencies PaymentFrequcey {get;set;}
-        public int InterestRate {get;set;}
+        public decimal InterestRate {get;set;}
         public int Spread {get;set;}
         public int CombinedGDS {get;set;}
         public int CombinedTDS {get;set;}
